parte_tecnica/04_api_exposure/requirements.txt
# Steel Price Predictor API - Requirements
# Python 3.9+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0

# ML/Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pytz==2023.3
